Move prover_dal types to basic_types::prover_dal module
